+++ src/sdk/main.ts
@@ -1,7 +1,12 @@
 /* eslint no-console: 0 */ // --> OFF
-export { whoami, whoamiRequiredParameters } from './v2/whoami'
-export { status } from './v3/status'
-export { IRepoFragment, repos } from './v2/repos'
 export { IRepoParameters } from './utils/models'
 
+// V2
+export { whoami, whoamiRequiredParameters } from './v2/whoami'
+export { IRepoFragment, repos } from './v2/repos'
+
+// V3
+export { status } from './v3/status'
 export { IRepoExRefactored, reposExtended } from './v3/repos_extended'
+export * from './v3/last'
+export * from './v3/defaultBranch'